Fixed exception cache not found
On fresh install, users would recieve an error saying the cache file could not be found.

+++ TcNo-Acc-Switcher-Server/Pages/General/GeneralFuncs.cs
@@ -10,6 +10,7 @@ namespace TcNo_Acc_Switcher.Pages.General
     {
         public static bool DeletedOutdatedFile(string filename)
         {
+            if (!File.Exists(filename)) return true;
             if (DateTime.Now.Subtract(File.GetLastWriteTime(filename)).Days <= 7) return false;
             File.Delete(filename);
             return true;
